Tentando resolver bug

+++ back/src/database/migrations/20201027192347_create_product.js
@@ -2,11 +2,11 @@
 exports.up = knex => knex.schema.createTable('products', table => {
     table.increments('id').primary()
     table.string('name').unique().notNullable();
-    table.decimal('preco').notNullable();
+    table.integer('preco').notNullable();
     table.integer('quantidade').notNullable();
-    table.decimal('precoUnitario').notNullable();
+    table.integer('precoUnitario').notNullable();
     table.string('fornecedor');
-    table.decimal('precoCusto');
+    table.integer('precoCusto');
 })
 
 exports.down = knex => knex.schema.dropTable('products');
